+++ fugue/src/test/java/io/atlassian/fugue/TrySuccessTest.java
@@ -14,7 +14,6 @@ import java.util.stream.Stream;
 import static java.util.function.Function.identity;
 import static java.util.stream.Collectors.toList;
 import static org.hamcrest.Matchers.contains;
-import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.notNullValue;
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertThat;
@@ -36,26 +35,26 @@ public class TrySuccessTest {
     throw new TestException();
   };
 
-  @Test public void isFailure()  {
+  @Test public void isFailure() {
     assertThat(t.isFailure(), is(false));
   }
 
-  @Test public void isSuccess()  {
+  @Test public void isSuccess() {
     assertThat(t.isSuccess(), is(true));
   }
 
-  @Test public void map()  {
+  @Test public void map() {
     assertThat(t.map(f).map(g), is(t));
   }
 
-  @Test public void mapThrowingFunctionRetunsFailure()  {
+  @Test public void mapThrowingFunctionRetunsFailure() {
 
     Try<Integer> result = t.map(fThrows).map(g);
 
     assertThat(result.isFailure(), is(true));
   }
 
-  @Test public void flatMap()  {
+  @Test public void flatMap() {
     Try<String> t2 = t.flatMap(i -> Checked.now(() -> fChecked.apply(i)));
 
     assertThat(t2, is(Checked.now(() -> "1")));
@@ -67,27 +66,27 @@ public class TrySuccessTest {
     t.flatMap(fTryThrows);
   }
 
-  @Test public void recover()  {
+  @Test public void recover() {
     assertThat(t.recover(e -> 1), is(t));
   }
 
-  @Test public void recoverExceptionType()  {
+  @Test public void recoverExceptionType() {
     assertThat(t.recover(Exception.class, e -> 1), is(t));
   }
 
-  @Test public void recoverWith()  {
+  @Test public void recoverWith() {
     assertThat(t.recoverWith(e -> Checked.now(() -> 1)), is(t));
   }
 
-  @Test public void recoverWithExceptionType()  {
+  @Test public void recoverWithExceptionType() {
     assertThat(t.recoverWith(Exception.class, e -> Checked.now(() -> 1)), is(t));
   }
 
-  @Test public void getOrElse()  {
+  @Test public void getOrElse() {
     assertThat(t.getOrElse(() -> 1), is(STARTING_VALUE));
   }
 
-  @Test public void fold()  {
+  @Test public void fold() {
     Integer i = t.fold(v -> {
       throw new RuntimeException();
     }, identity());
@@ -95,17 +94,17 @@ public class TrySuccessTest {
     assertThat(i, is(STARTING_VALUE));
   }
 
-  @Test public void foldPassedThrowingFunctionThrows()  {
+  @Test public void foldPassedThrowingFunctionThrows() {
     thrown.expect(TestException.class);
 
     t.fold(x -> "x", fThrows);
   }
 
-  @Test public void toEither()  {
+  @Test public void toEither() {
     assertThat(t.toEither(), is(Either.right(STARTING_VALUE)));
   }
 
-  @Test public void toOption()  {
+  @Test public void toOption() {
     assertThat(t.toOption(), is(Option.some(STARTING_VALUE)));
   }
 
@@ -130,10 +129,6 @@ public class TrySuccessTest {
     assertThat(stream.collect(toList()), contains(STARTING_VALUE));
   }
 
-  @Test public void iterable() {
-    assertThat(t, contains(STARTING_VALUE));
-  }
-
   @Test public void forEach() {
     final AtomicInteger invoked = new AtomicInteger(STARTING_VALUE);
     t.forEach(v -> invoked.set(ANOTHER_VALUE + v));
@@ -148,9 +143,7 @@ public class TrySuccessTest {
 
   @Test public void orElseSuccessSupplier() {
     final Try<Integer> orElse = t.orElse(() -> Try.successful(ANOTHER_VALUE));
-    assertThat(orElse, notNullValue());
-    assertThat(orElse.isSuccess(), is(true));
-    assertThat(orElse, contains(STARTING_VALUE));
+    assertThat(orElse, is(t));
   }
 
   @Test public void orElseFailureInstance() {
@@ -160,27 +153,21 @@ public class TrySuccessTest {
 
   @Test public void orElseFailureSupplier() {
     final Try<Integer> orElse = t.orElse(() -> Try.failure(new TestException()));
-    assertThat(orElse, notNullValue());
-    assertThat(orElse.isSuccess(), is(true));
-    assertThat(orElse, contains(STARTING_VALUE));
+    assertThat(orElse, is(t));
   }
 
   @Test public void filterTrue() {
-    final Try<Integer> orElse = t.filter(value -> Objects.equals(value, STARTING_VALUE));
-    assertThat(orElse, notNullValue());
-    assertThat(orElse.isSuccess(), is(true));
-    assertThat(orElse, contains(STARTING_VALUE));
+    final TestException testException = new TestException();
+    final Try<Integer> filter = t.filter(value -> Objects.equals(value, STARTING_VALUE), () -> testException);
+    assertThat(filter, is(t));
   }
 
   @Test public void filterFalse() {
-    final Try<Integer> orElse = t.filter(value -> !Objects.equals(value, STARTING_VALUE));
-    assertThat(orElse, notNullValue());
-    assertThat(orElse.isSuccess(), is(false));
-    assertThat(orElse.toEither().left().get(), instanceOf(NoSuchElementException.class));
+    final TestException testException = new TestException();
+    final Try<Integer> filter = t.filter(value -> !Objects.equals(value, STARTING_VALUE), () -> testException);
+    assertThat(filter, is(Try.failure(testException)));
   }
 
-  private class TestException extends RuntimeException {
-    private static final long serialVersionUID = 2348808619239132482L;
-  }
+  private class TestException extends RuntimeException {}
 
 }