+++ changelog.md
@@ -6,7 +6,7 @@ This project attempts to adhere to [Semantic Versioning](http://semver.org/).
 ## [4.7.0] - [unreleased]
 ### Added
 - added `TryMatchers` to complement current set of hamcrest matchers
-- Try will implement Iterable and includes additional methods:
+- Try includes additional methods:
     * filter
     * orElse
     * forEach